requirements.txt
altgraph==0.17.3
arrow==1.2.2
asgiref==3.7.2
attrs==22.1.0
aws-lambda-builders==1.18.0
aws-sam-cli==1.53.0
aws-sam-translator==1.46.0
binaryornot==0.4.4
boto3==1.24.51
botocore==1.27.51
brotlipy @ file:///home/conda/feedstock_root/build_artifacts/brotlipy_1648854164373/work
certifi==2022.6.15
cffi @ file:///home/conda/feedstock_root/build_artifacts/cffi_1636046055389/work
chardet==4.0.0
charset-normalizer @ file:///home/conda/feedstock_root/build_artifacts/charset-normalizer_1655906222726/work
chevron==0.14.0
click==7.1.2
colorama @ file:///home/conda/feedstock_root/build_artifacts/colorama_1655412516417/work
conda==4.13.0
conda-package-handling @ file:///home/conda/feedstock_root/build_artifacts/conda-package-handling_1649385069444/work
cookiecutter==2.1.1
cryptography @ file:///home/conda/feedstock_root/build_artifacts/cryptography_1652967134882/work
dateparser==1.1.1
defusedxml==0.7.1
Django==4.2.8
djangorestframework==3.15.1
djangorestframework-xml==2.0.0
docker==4.2.2
Flask==1.1.4
Flask-Cors==3.0.10
idna==2.10
itsdangerous==1.1.0
Jinja2==2.11.3
jinja2-time==0.2.0
jmespath==0.10.0
jsonschema==3.2.0
MarkupSafe==2.0.1
pycosat @ file:///home/conda/feedstock_root/build_artifacts/pycosat_1649384802055/work
pycparser @ file:///home/conda/feedstock_root/build_artifacts/pycparser_1636257122734/work
pyinstaller==5.6.2
pyinstaller-hooks-contrib==2022.12
pyOpenSSL @ file:///home/conda/feedstock_root/build_artifacts/pyopenssl_1643496850550/work
pyrsistent==0.18.1
PySocks @ file:///home/conda/feedstock_root/build_artifacts/pysocks_1648857263093/work
python-dateutil==2.8.2
python-slugify==6.1.2
pytz==2022.2.1
PyYAML==5.4.1
regex==2021.9.30
requests==2.25.1
ruamel-yaml-conda @ file:///home/conda/feedstock_root/build_artifacts/ruamel_yaml_1653464393916/work
s3transfer==0.6.0
serverlessrepo==0.1.10
six @ file:///home/conda/feedstock_root/build_artifacts/six_1620240208055/work
sqlparse==0.4.2
text-unidecode==1.3
tomlkit==0.7.2
tqdm @ file:///home/conda/feedstock_root/build_artifacts/tqdm_1649051611147/work
typing_extensions==4.9.0
tzlocal==3.0
urllib3 @ file:///home/conda/feedstock_root/build_artifacts/urllib3_1647489083693/work
watchdog==2.1.2
websocket-client==1.3.3
Werkzeug==1.0.1
